+++ src/com/intellij/idea/plugin/hybris/tools/ccv2/actions/CCv2BuildActions.kt
@@ -23,6 +23,7 @@ import com.intellij.idea.plugin.hybris.settings.CCv2Subscription
 import com.intellij.idea.plugin.hybris.tools.ccv2.CCv2Service
 import com.intellij.idea.plugin.hybris.tools.ccv2.dto.CCv2Build
 import com.intellij.idea.plugin.hybris.tools.ccv2.ui.CCv2CreateBuildDialog
+import com.intellij.idea.plugin.hybris.tools.ccv2.ui.CCv2DeployBuildDialog
 import com.intellij.idea.plugin.hybris.toolwindow.ccv2.CCv2Tab
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.DataKey
@@ -60,6 +61,21 @@ class CCv2RedoBuildAction(
     }
 }
 
+class CCv2DeployBuildAction(
+    private val subscription: CCv2Subscription,
+    private val build: CCv2Build
+) : AbstractCCv2Action(
+    tab = CCv2Tab.BUILDS,
+    text = "Deploy the Build",
+    icon = HybrisIcons.CCV2_BUILD_DEPLOY
+) {
+    override fun actionPerformed(e: AnActionEvent) {
+        val project = e.project ?: return
+
+        CCv2DeployBuildDialog(project, subscription, build).showAndGet()
+    }
+}
+
 class CCv2DeleteBuildAction(
     private val subscription: CCv2Subscription,
     private val build: CCv2Build
@@ -83,9 +99,11 @@ class CCv2DeleteBuildAction(
     }
 }
 
-class CCv2FetchBuildsAction : AbstractCCv2FetchAction(
+class CCv2FetchBuildsAction : AbstractCCv2FetchAction<CCv2Build>(
     tab = CCv2Tab.BUILDS,
     text = "Fetch Builds",
     icon = HybrisIcons.CCV2_FETCH,
-    fetch = { project, subscriptions, onStartCallback, onCompleteCallback -> CCv2Service.getInstance(project).fetchBuilds(subscriptions, onStartCallback, onCompleteCallback) }
+    fetch = { project, subscriptions, onStartCallback, onCompleteCallback ->
+        CCv2Service.getInstance(project).fetchBuilds(subscriptions, onStartCallback, onCompleteCallback)
+    }
 )