+++ src/com/intellij/idea/plugin/hybris/tools/ccv2/actions/CCv2DeploymentActions.kt
@@ -20,12 +20,15 @@ package com.intellij.idea.plugin.hybris.tools.ccv2.actions
 
 import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
 import com.intellij.idea.plugin.hybris.tools.ccv2.CCv2Service
+import com.intellij.idea.plugin.hybris.tools.ccv2.dto.CCv2Deployment
 import com.intellij.idea.plugin.hybris.toolwindow.ccv2.CCv2Tab
 
-class CCv2FetchDeploymentsAction : AbstractCCv2FetchAction(
+class CCv2FetchDeploymentsAction : AbstractCCv2FetchAction<CCv2Deployment>(
     tab = CCv2Tab.DEPLOYMENTS,
     text = "Fetch Deployments",
     icon = HybrisIcons.CCV2_FETCH,
-    fetch = { project, subscriptions, onStartCallback, onCompleteCallback -> CCv2Service.getInstance(project).fetchDeployments(subscriptions, onStartCallback, onCompleteCallback) }
+    fetch = { project, subscriptions, onStartCallback, onCompleteCallback ->
+        CCv2Service.getInstance(project).fetchDeployments(subscriptions, onStartCallback, onCompleteCallback)
+    }
 )