+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - Fetch CCv2 deployments via `Native` integration [#1073](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1073)
 - Schedule new build via `Native` integration [#1074](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1074)
 - Delete existing build via `Native` integration [#1075](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1075)
+- Schedule deployment for existing build [#1076](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1076)
 
 ### `Type System` enhancements
 - Show Item line marker for jalo classes [#1067](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1067)